removed layout for passing state to children, improved preloading of images
